examples/imx7_colibri_m4: select correct clock for GPTA
GPTA is supposed to run directly of the 24MHz OSC. Fix a type which
update GPTB's root clock instead of GPTA.

Note: GPTB seems to rely on the System PLL PFD0, which seems to be
influenced by Linux' SDHC driver. Freezes of both systems have been
observed. Disabling the SDHC driver on the Linux side or using the
24MHz clock source for GPTB seem to avoid freezes.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>