Fix UART examples for Colibri iMX7
The uart_polling and uart_interrupt examples were originally written for a UART used in DCE mode. The Colibri iMX7 is designed to use the UART in DTE mode. The signals are swapped between DTE and DCE (Rx <-> Tx, and control signals).
In order to use the correct pin direction, the bit DCEDTE in the UARTx_UFCR register needs to be configured accordingly.

Signed-off-by: Andy Kiser <andy.kiser@gmx.ch>