increase memory in DDR linker file
Increase the memory (m_data) area in the DDR linker file. Also use
the code bus for the code area instead of the system bus. This
aligns with the _tcm.ld file and should improve performance.

Note: To get more heap space in FreeRTOS configTOTAL_HEAP_SIZE in
FreeRTOSConfig.h needs to be adjusted as well.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>